requirements.txt
python_telegram_bot==13.3
Flask==1.1.2
